WIP: Add app menu for some common things.
This doesn't yet implement the suggestion by Allan, but it's a start. Some
more work needs to be done regarding Find Remote & Sync (since it's an
optional at compile time menu.
Also, view isn't yet in the AppMenu because I think it would be better as a UI
switch/button.
    Also needs some testing (I've seen a crash).

https://bugzilla.gnome.org/show_bug.cgi?id=674979